+++ lib/multi_qr_tracker.dart
@@ -17,5 +17,7 @@ library;
 
 export 'src/models/qr_code_info.dart';
 export 'src/models/qr_tracker_camera_orientation.dart';
+export 'src/models/torch_mode.dart';
 export 'src/multi_qr_tracker_view.dart';
 export 'src/widgets/qr_border_painter.dart';
+export 'src/widgets/torch_button.dart';